+++ askbot/skins/common/media/js/user.js
@@ -18,7 +18,7 @@ $(document).ready(function(){
     };
 
     var submit = function(id_list, elements, action_type){
-        if (action_type == 'delete' || action_type == 'mark_new' || action_type == 'mark_seen'){
+        if (action_type == 'delete' || action_type == 'mark_new' || action_type == 'mark_seen' || action_type == 'remove_flag' || action_type == 'close' || action_type == 'delete_post'){
             $.ajax({
                 type: 'POST',
                 cache: false,
@@ -27,7 +27,7 @@ $(document).ready(function(){
                 url: askbot['urls']['manageInbox'],
                 success: function(response_data){
                     if (response_data['success'] === true){
-                        if (action_type == 'delete'){
+                        if (action_type == 'delete' || action_type == 'remove_flag' || action_type == 'close' || action_type == 'delete_post'){
                             elements.remove();
                         }
                         else if (action_type == 'mark_new'){
@@ -61,11 +61,35 @@ $(document).ready(function(){
                 return;
             }
         }
+        if (action_type == 'close'){
+            msg = ngettext('Close this entry?',
+                    'Close these entries?', data['id_list'].length);
+            if (confirm(msg) === false){
+                return;
+            }
+        }
+        if (action_type == 'remove_flag'){
+            msg = ngettext('Remove all flags on this entry?',
+                    'Remove all flags on these entries?', data['id_list'].length);
+            if (confirm(msg) === false){
+                return;
+            }
+        }
+        if (action_type == 'delete_post'){
+            msg = ngettext('Delete this entry?',
+                    'Delete these entries?', data['id_list'].length);
+            if (confirm(msg) === false){
+                return;
+            }
+        }
         submit(data['id_list'], data['elements'], action_type);
     };
     setupButtonEventHandlers($('#re_mark_seen'), function(){startAction('mark_seen')});
     setupButtonEventHandlers($('#re_mark_new'), function(){startAction('mark_new')});
     setupButtonEventHandlers($('#re_dismiss'), function(){startAction('delete')});
+    setupButtonEventHandlers($('#re_remove_flag'), function(){startAction('remove_flag')});
+    setupButtonEventHandlers($('#re_close'), function(){startAction('close')});
+    setupButtonEventHandlers($('#re_delete_post'), function(){startAction('delete_post')});
     setupButtonEventHandlers(
                     $('#sel_all'),
                     function(){
@@ -90,6 +114,16 @@ $(document).ready(function(){
                     function(){
                         setCheckBoxesIn('#responses .new', false);
                         setCheckBoxesIn('#responses .seen', false);
+                    }
+    );
+
+    setupButtonEventHandlers($('.re_expand'),
+                    function(e){
+                        e.preventDefault();
+                        var re_snippet = $(this).find(".re_snippet:first")
+                        var re_content = $(this).find(".re_content:first")
+                        $(re_snippet).slideToggle();
+                        $(re_content).slideToggle();
                     }
     );
 });